Improve connection handling
Since the connection could be coming from a custom factory, there are
some assumptions that we cannot make as if it was an Active Record
connection from the current thread. So, we have to assume calling the
factory will open a new connection each time (or borrow it from the
pool) and that we need to close it when we are done.