Accept an optional IO stream for output instead of the DB connection
The idea is to allow the caller to write a SQL file with the data to be
loaded. This could be useful if the data needs to be loaded multiple
times. Generating the SQL statements themselves is time consuming, so by
saving a SQL file and reusing it, a significant amount of time could be
saved.