Switch from accepting an IO stream to a file parameter instead
The IO stream option had the problem of not knowing the context on how
it would be used. So, to make that work for COPY it was providing a
PROGRAM that echoed the batch of data. However this approach breaks if
the echo is large enough which would vary by the width of the data
along with the batch size.

Instead, now we will accept a file option and create the files on disk
based on the filename given. Instead of COPY, we use \COPY and create
the corresponding CSV data files along side the SQL script file. The
assumption being that it will be executed via psql on the client.