Adjust the statement_timeout when the output is a script file.
For postgres, we want the script file to adjust the statement timeout at
the very top of the script and put it back to the original value at the
end of the script.

This is independent from needing to adjust the statement timeout on the
connection itself which is still needed for the read operations that
take place while generating the data sets.