Change output behavior to be in addition to loading the database.
The output file will be optionally written but the database will always
be loaded upon execution. The problem with trying to only write a SQL
script without loading the database is that the related data on a
belongs_to depends on parent tables having been already populated. So
the user would need to separate their scripts and make sure they are run
in the correct order.

Instead, we can load the data to begin with and also write the SQL
script. So from that point forward the user can use the SQL scripts to
reload the same data later on and save time that way.