Fix bug with cycling on two foreign keys in a unique index
If two belongs_to columns are part of a unique index then the data
providers would cycle through the existing IDs in order. When reaching
the end of the list they will generate the same pattern again since they
will continue to cycle in the same order.

A simple solution to this (given we retry rows) is to shuffle the list
in each cycle so retrying will have an opportunity of coming up with
unique values.